Codex - Issue #60 - Implement work history page
Extract tag normalization and featured selection from ProjectRepository.
Fix normalized tag migration deduplication and work-history role ordering.
