+++ ProjectPortfolio2026/ProjectPortfolio2026.Server/Program.cs
@@ -72,6 +72,8 @@ builder.Services
 builder.Services.AddAuthorization();
 builder.Services.AddScoped<IPortfolioProfileRepository, PortfolioProfileRepository>();
 builder.Services.AddScoped<IPortfolioLinkFormatter, PortfolioLinkFormatter>();
+builder.Services.AddScoped<IProjectTagNormalizer, ProjectTagNormalizer>();
+builder.Services.AddScoped<IFeaturedProjectSelector, FeaturedProjectSelector>();
 builder.Services.AddScoped<IProjectRepository, ProjectRepository>();
 builder.Services.AddScoped<IEmployerRepository, EmployerRepository>();
 builder.Services.AddScoped<ICurrentUserAccessor, CurrentUserAccessor>();
