+++ ProjectPortfolio2026/ProjectPortfolio2026.Server/Services/Interfaces/IProjectTagNormalizer.cs
@@ -1,0 +1,8 @@
+using ProjectPortfolio2026.Server.Domain.Projects;
+
+namespace ProjectPortfolio2026.Server.Services.Interfaces;
+
+public interface IProjectTagNormalizer
+{
+    Task NormalizeAsync(Project project, CancellationToken cancellationToken = default);
+}
